refactor: create api folder to replace previous store folder
