+++ src/api/index.js
@@ -1,0 +1,5 @@
+export { default as contentful } from "./contentful";
+export * from "./contentful";
+
+export { default as koala } from "./koala";
+export * from "./koala";
