+++ src/store.js
@@ -1,8 +1,38 @@
-import { configureStore } from "@reduxjs/toolkit";
+import { createSlice, configureStore } from '@reduxjs/toolkit';
 
-import koala from "./store/koala";
-import contentful from "./store/contentful";
-import state from "./store/state";
+import { koala, contentful } from './api';
+
+const state = createSlice({
+  name: 'state',
+  initialState: {
+    activityIndex: 0,
+    adIndex: 0,
+    boardMessageIndex: 0,
+    quotesIndices: []
+  },
+  reducers: {
+    incrementActivityIndex(state) {
+      state.activityIndex++;
+    },
+    incrementAdIndex(state) {
+      state.adIndex++;
+    },
+    incrementBoardMessageIndex(state) {
+      state.boardMessageIndex++;
+    },
+    shiftQuote(state, action) {
+      if (state.quotesIndices.length === 1) {
+        state.quotesIndices = new Array(action.payload)
+          .fill(0)
+          .map((_, i) => i);
+      } else {
+        state.quotesIndices.shift();
+      }
+    }
+  }
+});
+
+export const { incrementActivityIndex, incrementAdIndex, incrementBoardMessageIndex, shiftQuote } = state.actions;
 
 const store = configureStore({
   reducer: {
